fix(core): 修复中等问题 - 线程安全和精度问题
修复内容：
1. EntityClearerService 使用 ConcurrentHashMap.newKeySet() 替代 mutableSetOf()
2. RealWorldStorage 使用容差比较替代浮点数直接相等比较
3. FoodCorrosionEngine.finalizeStorageExit 使用浮点数除法避免精度丢失
4. RealWorldService.init 使用 firstOrNull() 避免异常
5. WeatherEngine.init 接受 Long 类型种子避免溢出

构建验证：BUILD SUCCESSFUL

+++ project/core/src/main/kotlin/com/pixlehavencore/feature/realworld/weather/ChunkWeatherEngine.kt
@@ -14,8 +14,8 @@ object ChunkWeatherEngine {
     val isInitialized: Boolean
         get() = rainNoise != null
 
-    fun init(worldSeed: Int, frequency: Float = 0.015f) {
-        rainNoise = FastNoiseLite(worldSeed).apply {
+    fun init(worldSeed: Long, frequency: Float = 0.015f) {
+        rainNoise = FastNoiseLite(worldSeed.toInt()).apply {
             setFrequency(frequency)
         }
     }
